// src/shared/api/models/getApiSalesListSalesType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

export type GetApiSalesListSalesType = (typeof GetApiSalesListSalesType)[keyof typeof GetApiSalesListSalesType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiSalesListSalesType = {
    SALES: 'SALES',
    RETURN: 'RETURN',
    NOSHOW: 'NOSHOW',
} as const;