// src/shared/api/models/getApiSalesListSortBy.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

export type GetApiSalesListSortBy = (typeof GetApiSalesListSortBy)[keyof typeof GetApiSalesListSortBy];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiSalesListSortBy = {
    ALL: 'ALL',
    CASH: 'CASH',
    CARD: 'CARD',
    GIFT_CARD: 'GIFT_CARD',
    CHEQUE: 'CHEQUE',
    ECOMMERCE: 'ECOMMERCE',
    OTHER: 'OTHER',
} as const;