// src/shared/api/models/getApiSalesListSortOrder.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

export type GetApiSalesListSortOrder = (typeof GetApiSalesListSortOrder)[keyof typeof GetApiSalesListSortOrder];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiSalesListSortOrder = {
    asc: 'asc',
    desc: 'desc',
} as const;